stmt2unit - mark things on module exit. If the "END" statement is generic we need to go through the generic_exit to detect if the end applies to a module - NOTE this may turn out also to be necessary for other generic END cases; fix up debug statements - also NOTE using "self" as a parameter name for a stand-along method is misleading;
symtab.py - mark things referenced and optionally also global - this should use the same logic to descend into the types for generic interfaces or else we will miss those types when we want to do generic resultion later.  The descendence logic is the same for the reference marker and the global marker;   ALSO drop the implicit list for now from the symbol table debug - this just clutters the output.